refactor: 컴포넌트 디렉토리 경로 수정
